backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
apscheduler==3.10.4
google-generativeai==0.3.2
gunicorn==21.2.0
werkzeug==3.0.1
sqlalchemy==2.0.25
jinja2==3.1.3
click==8.1.7
itsdangerous==2.1.2
markupsafe==2.1.5
